requirements.txt
django
easy-thumbnails
djangorestframework
stripe
django-paypal
django-ckeditor
django-markdown
